app-backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-cloud-firestore
google-cloud-storage
google-genai
pydantic
python-dotenv
